Don't pass Request classes into services
